# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Messenger backend api",
        "description": "Messenger backend",
        "contact": {
            "email": "tcom.developer@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://nighthub.test/api",
            "description": "Dev server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "user name",
                                        "type": "string",
                                        "default": "newuser"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "default": "newuser@mail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "default": "Secret123#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/resend-verify-email": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "resend verify email",
                "operationId": "39227350a2b92d59775b309975bce2d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "default": "resend@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verify email sent"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/verify-email": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify email address",
                "operationId": "6902a4b313dbb202a9d4592dc5acd616",
                "parameters": [
                    {
                        "name": "verify_token",
                        "in": "query",
                        "description": "Email verification token from email link",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to frontend login page"
                    },
                    "400": {
                        "description": "Invalid or expired token"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "default": "milos@mail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "default": "milos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Korisnik prijavljen"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/auth/whoami": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Podaci o korisniku na osnovu tokena",
                "operationId": "aa7bc66822f59c40cade93c5e4c3b41e",
                "responses": {
                    "200": {
                        "description": "Podaci o korisniku vraćeni na front"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Odjava sa sistema",
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Korisnik odjavljen iz sistema"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/search": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Search user by name or email",
                "operationId": "7495095e5b6c23625c419daaa0200f1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "search"
                                ],
                                "properties": {
                                    "search": {
                                        "description": "search string",
                                        "type": "string",
                                        "default": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/show/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Find user by ID",
                "operationId": "2359a38d992595bc4001f41c0b92d89d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/edit": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Find user by auth token",
                "operationId": "fef4feace1e1197ecf512a20eb5bf906",
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/update": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update user data",
                "operationId": "25fa6e59311f61c2be4cca80c602dde3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "about",
                                    "avatar"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "about": {
                                        "description": "About user",
                                        "type": "string",
                                        "example": "Software developer from Belgrade"
                                    },
                                    "avatar": {
                                        "description": "User avatar file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/forgot-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Send reset password to email",
                "operationId": "5730d16a34a3a9c69039ba5aa91552f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "milos.glogovac@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset password link sent"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/change-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change user password",
                "operationId": "48fbc004e5b7e15198e50b6409fbd7ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "about",
                                    "avatar"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "Password123#"
                                    },
                                    "password2": {
                                        "description": "User password repeated",
                                        "type": "string",
                                        "example": "Password123#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Controller": {
                "format": "https"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authorize with bearer token",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}